pyproject.toml
[tool.poetry]
name = "streamcontrol"
version = "0.1.0"
description = ""
authors = ["chalupaul <me@chalupaul.com>"]

[tool.poetry.dependencies]
python = "^3.9"
obs-websocket-py = "^0.5.3"
python-dotenv = "^0.19.2"
structlog = "^21.5.0"
colorama = "^0.4.4"
astral = "^2.2"
aiocron = "^1.8"
Jinja2 = "^3.0.3"

[tool.poetry.dev-dependencies]
pytest = "^7.0.1"
pytest-mock = "^3.7.0"
pytest-cov = "^3.0.0"
taskipy = "^1.10.1"
black = "^22.1.0"
isort = "^5.10.1"
flake8 = "^4.0.1"
bandit = "^1.7.4"
doit = "^0.34.2"
toml = "^0.10.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"

[tool.black]
line-length = 88
include = '\.pyi?$'
exclude = '''
/(
  | \.git
  | .venv
  | .coverage
  | __pycache__
)/
'''

[tool.poetry.scripts]
app = "streamcontrol.main:main"

[tool.taskipy.tasks]
tests = """pytest --cov-report term-missing --cov=streamcontrol tests"""